Add a UDP service port for quic
Recent storj storage node releases have added a new UDP port to use the
QUIC protocal in addition to the standard TCP port
(https://forum.storj.io/t/experimenting-with-udp-based-protocols/11545).
This commit will add the new UDP service to the Chart, allowing it to be
exposed.

As part of this addition, the service logic in `values.yaml` and
`service.yaml` was updated to move all services under a common `service`
dictionary and iterate over them. The `nodeStats` object no longer
exists and was moved to `service.stats`.